Create-abstract-class-for-issue#30
